provider/maas: remove retry logic from bridge script
Remove the retry logic on various commands that have been introduced
recently for migrating the IP addresss and the routes from the parent
device to the bridge device.

I was seeing a lot of:

  RTNETLINK: file exists

when boostrapping.

Removing addresses from the parent is also done using a single call:

  $ ip addr flush dev eth0

and we now only run:

  $ ip route replace {route} dev {bridge}

if there is a 'default' route for the {parent}.

Previously we were doing 'ip route replace' for each bridge and it was
that command that was generating the 'RTNETLINK: file exists' messages
because the following (which we issue) automatically adds a route for
the device:

  $ ip addr add dev {bridge} {address}

Fixes [LP:#1588446](https://bugs.launchpad.net/juju-core/+bug/1588446)