state/presence: fix error handling in sync
Updates LP 1588574

While investigating LP 1588574 I came across this piece of error
handling logic that was addedin back in 2012 in 93d52ed7.

The logic was non sensical; if err was not nil, but was MgoNotFound then
wrap it and return it, _otherwise_ continue on. I'm assuming that this
should actually be err was not nil and not MgoNotFound then return the
error.