all: don't leak pingers in tests
Updates LP 1590161

The fixtures for the many tests were starting pingers and loosing
track of them; assuming that they would die eventually when mongodb
shuts down ... which is true, except their death was not as private
an afair as previously assumed.

Editorial: the root cause of this problem is the mgo driver's copying
semantics -- rather than obtaining a connection to the database, you
copy a collection which is assumed to be already connected. Because this
copy operation does not return an error, the driver has forced itself
into using a panic.

Additionally the SetAgentPresence API on various 'present' things comes
with a footgun of not making clear just how important it is to keep
track of the thing that is tracking presence. The name of that API,
`SetAgentPresence` lulls the reader into thinking it's some kind of
boolean setter "yes, I'd like to track the presence of this thing."
There is little clue that calling that method returns a resource that
if not handled carefully will cause your program to panic randomly.