Clean up interface migration binary uploads
This change removes unnecessary indirection in
migration.UploadBinaries. Instead of dealing with api.Connection and
conversion helpers the desired interfaces are accepted directly. This
cleans up the testing and implementation of both UploadBinaries and
where it is used in the migrationmaster worker.

The NewUploadBinariesConfig function has been removed now that
UploadBinariesConfig is easier to construct. It's good to be explicit
anyway.