worker/catacomb: do not decrement catacomb.wg until worker has stopped
This change is a prerequisite of #5564, LP 1590161

Prior to this change, with #5564 applied, the cmd/jujud/agent tests
would reliably cause a 'session already closed' panic because the
workers owned by the catacomb had not yet stopped.

This change alters the catacomb.add logic to use the worker.Stop helper
that ensures the worker has stopped before decrementing catacomb.wg.
In turn, assuming that the catacomb's owner waits until its tomb is Dead,
this ensures that the catacomb is not considered dead until all its
workers are fully stopped.

With this change in place, the cmd/jujud/agent tests pass reliably.