Remove the User and ModelUser(s) seperately
We had them lumped into the same mgo txn, but if there is no User then
we fail in the transaction and get a less than useless error. So now
we remove the user, unless it doesn't exist, then try removing the
ModelUser(s). If there's no ModelUser we don't care.