worker/firewaller: register machined with firewaller's catacomb
Updates 1590161

Firewaller.startMachine registers the unitd with the containing
firewaller's catacomb, but the machined which mangers units is not
tracked. This machined will die when any of its component units dies,
but its lifetime is not bound to the containing firewaller's catacomb,
so it can remain alive when the firewaller is considered stopped.

This can lead to panics when pingers for the machine that the machined
is responsible for have not died by the time the test is being torn
down.