Fix lp:1590960 - deploy --to lxd
`juju deploy <charm> --to <container-type>` wouldn't create a new parent
machine for the container - it would create the unit but no machines for
it. Change unit assignment to be consistent with the behaviour of `juju
add-machine <container-type>` - it will create a new machine and then
deploy the charm to a container on that machine.

https://bugs.launchpad.net/juju-core/+bug/1590960