state: Introduce notifyCollWatcher
The rebootWatcher, migrationStatusWatcher and cleanupWatcher were
slight variations of the same watcher. The implementations have been
extracted into a single notifyCollWatcher which handles all of them.

This is a precursor to a new watcher for migration minion reports
which will also be based on notifyCollWatcher.

Drive-by: the unused error return on a number of Watch* methods was
removed (leading to additional fallout in the apiserver).

Drive-by: removed silly collectionWatcher StringWatcher interface
check. newcollectionWatcher ensure this already by returning a
collectionWatcher as a StringWatcher.