worker/migrationminion: Report to master on SUCCESS
The migrationminion now reports to the master when it has processed
the SUCCESS phase. The master will wait for reports from all minions
before moving on to the next phase.

This change also cleans up handling of phases which aren't interesting
to the migration minion. The Guard (aka Fortress) gets unlocked for
all such phases.

Some unimplemented and untested handling for QUIESCE and VALIDATION
have been removed.