Store workload version in status
Rather than storing it as an attribute of the unit doc. We want to avoid
adding new fields to existing documents - each one means that there are
more possible things that could change that would notify all of the
watchers hanging off the document, increasing the likelihood of
unnecessary message storms.