bootstrap: interpret provider/<thing> as endpoint
When bootstrapping and specifying a cloud *type*,
e.g. "manual" or "maas", then the bit after /
will now be interpreted as the cloud endpoint
rather than as a region name. This means that
it won't show up in the controller or model
details as a region anymore, but also means
we can simplify some logic: no need to deal with
the possibility of a region that the cloud does
not support.

Fixes https://bugs.launchpad.net/juju-core/+bug/1593033