provider/dummy: do not close previous ops listener
Fixes LP 1588143

This PR removes the logic inside dummy.Listen that closes the previous
ops listener channel when being _replaced_. By not closing we avoid the
race condition with another goroutine which is trying to send a final
ops message. This change had no effect on the test cases and passes all
my stress tests.

Therefore this test not only fixes the bug, but increases test coverage :)