juju: add support for redirection when connecting to a model
We are going to have controllers which multiplex other
controllers. In such a case, we don't want to require
all the network traffic for the models that the sub-controllers
control to pass through the top level hub, so allow
a controller to return a "redirect required" error to cause
a client to connect to the model API in the sub-controller
(or some other server) directly.

There is more work to be done after this lands. In particular,
we should cache the API addresses that we were redirected
to and associate them with the model, so we can connect
directly on subsequent API opens.