Replace agent machine lock with juju/mutex.
This work completely removes all use of fslock from the juju core codebase.

The uniter hook execution log is now a named juju/mutex. The name is defined as a constant in the agent package, and passed to the workers that need it as manifold configuration.

The jujuclient store fslock is replaced with a global mutex. Since this mutex is only guarding around very bried read or write operations to a file, a system lock will have no appreciable difference to a user. A follow up branch could fix this should it become a real problem (which I don't think it will).