Replace hook execution fslock with mutex implementation.
This work completely removes all use of fslock from the juju core codebase.

The uniter hook execution lock is now a named juju/mutex. The name is defined as a constant in the agent package, and passed to the workers that need it as manifold configuration for those that have manifolds, or as params to those that don't.