bootstrap: dirty hack to support gce jsonfile
When bootstrapping, we check to see if the auth-type
is "jsonfile", and the provider is "gce". If so, we
read the credentials and transform to the oauth2
credential type.

This is a hack to get things working in beta10. We
will be making major changes following beta10 that
mean this bug will be fixed properly.

Fixes https://bugs.launchpad.net/juju-core/+bug/1593761