worker/migrationmaster: Wait for minions at SUCCESS
This change implements the infrastructure for the migration master to
wait for minions to report success or failure for the current
migration phase. This is used to wait at the SUCCESS phase, allowing
the removal of the temporary sleep that had been used in lieu of
actual synchronisation.

A lot of cleanup has been done in the worker tests to reduce
duplication and remove unnecessary setup related "noise".