Only require "authorized-keys" at bootstrap
We make authorized-keys an optional configuration
attribute, except at bootstrap time. This is already
enforced by the bootstrap code, so we're really just
making it an optional config attribute.

Also, move the handling of authorized-keys-path out
of environs/config, and into the client/CLI code.
That is the only place where it makes sense.

Finally, we add authorized-keys to the config when
creating a new model from the CLI, if it is possible
to do so. This means that we no longer need the
special-cased authorized-keys copying for models
created by controller admins.