Update catacomb to gracefully handle panic().
Without this patch, if a worker that a catacomb was managing would panic(), it
would end up getting hung in defer catacomb.wg.Wait(), because the Work()
didn't return an error that was being passed into Kill(), that meant the
catacomb's monitoring goroutine never went to Dying(), so the WG was never
decremented. (And it appears that because the catacomb was hung in a defer,
the panic could not escape and crash the rest of the system.)

This changes it so that panic in a work function just gets turned into
a normal error, so that we can continue to clean up and error out.