Use application-version-set instead of workload-
Renamed to be clearer, made the help text much more informative.

Removed workload/application-version-get because it's not really needed. As
implemented its behaviour would be confusing - a version set on one unit
wouldn't be returned by the get on another unit (since under the hood
the unit versions are stored separately). Remove it rather than doing
the extra work to make it always return the application-wide version.