cmd/jujud/agent/model: state-cleaner runs more
This *may* fix or mitigate lp:1591387, but is more correct than before
regardless: there's no good reason to stop running cleanups just because
a model is dead, and failing to do so is surprising from the perspective
of someone writing state code and expecting cleanups to run
automatically.

Also includes drive-by refactorings/cleanups that seemed sensible while
hunting for the bug.