Fixes a data-race in audit/logger observers
Fixes: https://bugs.launchpad.net/juju-core/+bug/1598063

Previously, Observers were called into at both the API server level and the RPC request level. This was causing issues as state is intended to be stored within the observer instead of being passed in. The crux of the problem was that the Observer was doing too much.

Observers now instead know how to create rpc.Observers which are then created for each RPC request. This isolates state, and the types which observe events are now split along RPC and API server delineations.