status: application workload-version -> version
In the `juju status` output, renamed workload-version to just version at
the application level and removed the unit-level workload-version
display. This is clearer and more how we want to present the
information, even though we're still capturing it at the unit
level. (From discussion and demoing to rick_h and jam.)