status: add app version to tabular format
Display the application version in the tabular status output. Limit the
length of the version displayed to 7 characters to prevent the table
from being blown out by the charm-supplied information. That's long
enough for most 1.2.3-style version numbers and most Git hashes. If the
version is truncated by this restriction the full version string will
still be visible in the yaml-formatted status. (From discussion with jam
and rick_h.)