juju status: fix flickering peer relation name
Fixes https://bugs.launchpad.net/juju-core/+bug/1598897

If you deployed a charm with a peer relation where the ends are named
differently, which name was displayed for the relation depended on the
map ordering. This meant that running `watch juju status` would show the
name flickering between the different options.

Sort the relation types before combining and displaying them so we pick
the relation name consistently.