all: stop abusing SSH bootstrap timeouts
Rather than using SSH bootstrap timeouts
directly in places where they weren't
intended for use, we now thread the values
through. In particular, we send timeouts
through to the bootstrap "finalizer", and
we add a --timeout flag to the jujud
bootsrap-state "agent".

There is a tangentially related cleanup to
do with instance config: we extract the
bootstrap-specific code from
instancecfg.FinishInstanceConfig, and move
it into the environs/bootstrap package.
This will enable further changes to
separate bootstrap and controller config.