provider/maas: add missing auto stanza
Ensure that the sub-interface has an 'auto' stanza if the original
declaration had an 'auto' stanza.

This is particularly important if the sub-interface has configuration
options (e.g., 'mtu') because our subsequent invocation of 'ifup -a'
only applies to those interfaces that are marked 'auto'. If 'auto' is
missing, then any MTU value would not be applied.

Fixes [LP:#1596597](https://bugs.launchpad.net/juju-core/+bug/1596597)