lp1577949: Don't set password if unit installed
When we switched to building with go 1.6 for all
agents, there was a problem where windows unit agents
could not upgrade.  The issue was that the windows
services for the unit agents did not show up in results
from ListServices immediately after the upgrade.

This resulted in the machine agent thinking the units
were not deployed and it would reset the password for the
unit agent, but then fail to deploy the unit because it
was already installed.  The unit agent would then fail to
log in as its password had been reset.

For 1.25.6, this patch re-checks that the service is
not installed immediately before resetting the password
for the unit.  This change has been tested in CI and I
verified that the problem recreated and this patch fixed
the issue.

This patch does not eliminate the timing window where
this occurs, but shortens it significantly.

For later releases, we need to determine why ListServices
is racy and has a window where it doesn't return the juju
services that are installed.