state: Ensure indexes are used when tailing logs
MongoDB has a 32MB limit when sorting without an index. Juju's logs
can easily exceed this limit so it is important to ensure that indexes
are used when sorting logs.

A recent change added the _id column to the sort fields causing
MongoDB to no longer uses indexes, breaking logging when there are any
signficant amount of logs present. This change splits sorting into 2
steps so that indexes can be used.

Fixes https://pad.tv/1590605