provider/maas: remove MTU option from the bridge options
The MTU statement must be on the interface to work and not on the bridge
itself; bridges inherit the lowest MTU of all sub-interfaces. Further to
this, ifupdown seems to bug out and always fails to add the default
route when an MTU is set on the bridge. Leaving the MTU option on the
bridge completely breaks the network on reboot.

This commit also removes:

  address
  gateway
  netmask
  dns-nameservers
  dns-search
  dns-sortlist

from the sub-interfaces, leaving them on the bridge interface only.

Fixes [LP:#1584616](https://bugs.launchpad.net/juju-core/+bug/1584616)