environs/bootstrap: fix test on windows
We were using the result of NormalizePath in an
error message, and expecting the path to be the
same on Linux and Windows. NormalizePath converts
slashes to native dir separators.

Rather than changing the test, the implementation
is changed to use the input to NormalizePath instead
of the output, when formatting the error message.

Fixes https://bugs.launchpad.net/juju-core/+bug/1603208