Introduced database access for auditing
This uses the ubiquitous pipeline[1] pattern and returns a channel upon
which records of FetchedAuditEntry or errors will be sent. The reader
will re-open an Iterator should it be necessary to provide a continuous
stream of records to listeners. All operations can be interrupted and
cleanly torn down by closing the done channel passed in.

[1] - https://blog.golang.org/pipelines