Introduced auditing-enabled config flag
Previous to this, auditing of an environment (i.e. all controllers & models) was automatic and unavoidable. This can lead to issues as all API events are logged to Mongo. This provides a config value for the controller (and thus the environment) that enables users to toggle this on and off. The default is for it to be switched off.

I also did a drive-by on removing usages of fmt.Errorf in favor of github.com/juju/errors.Errorf.