Fix the reboot worker tests.
Due to the behaviour of the reboot worker, the tests need to have different machine lock names, otherwise they will block on windows.
The golang garbabe collector means that this isn't a problem on linux as when the old lock is cleaned up, it releases the domain socket,
but on windows, the lock is an external semaphore that relies on the process dying to release it.