worker/uniter/charm: make test reliable
Removed dead code intended to block an HTTP request to a server that's
not actually being used by the SUT. Replaced concurrent abort check
(that we don't actually have a way to block) with synchronous abort
check against a pre-closed channel, which still demonstrates that the
abort chan *is* passed to the layer below, which should honour it.

Technically I suspect this is a mitigation not a *fix*, because it's
still *possible* for the download to complete in the gap between Start
and Wait in downloader/downloader.go; but the timing is certainly much
tighter, and I think that's progress worth pushing.

fixes lp:1604006