state: refactor state.Policy
The primary change in this diff is tidying
up the state.Policy interface, so that:
 - there are no assumptions built into the
   interface about how to construct an
   Environ
 - the environs package no longer has a
   dependency on state

There is a new package, state/stateenvirons,
which contains the old environs/statepolicy.go.
There are some additional helpers in the same
package to:
 - convert a *state.State into an
   environs.EnvironConfigGetter. Currently
   *state.State implements that interface,
   but it soon won't
 - construct an Environ given a *state.State,
   and a function to construct an Environ
   given an environs.EnvironConfigGetter

There are some changes scattered, mostly
throughout apiserver, to make use of the
EnvironConfigGetter interface more in tests.

This is yet another refactoring along the
path to untangling model config from cloud
and credential definitions, and updating
the Environ interface to match.