environs: refactor PrepareForBootstrap
We move PrepareForBootstrap off EnvironProvider,
onto Environ, and change its purpose. Once upon
a time it had a dual purpose: to add things to
the initial config, and to perform possibly-interactive
pre-bootstrap environment initialisation. We now
have the EnvironProvider.BootstrapConfig method
for initialising config; and the pre-bootstrap
init is best situated next to Bootstrap, on Environ.

Apart from clarifying the purpose of the method,
this change means that there is now exactly one
provider method for constructing an Environ, and that
is EnvironProvider.Open.

There is a small change to jujuclient related to
bootstrapping: to store the cloud type in bootstrap
config. This will become necessary when we remove
"type" from model config.