Update the mgo patch to fix max-retry bug
The upsert retry patch for mgo had a bug that would prevent the error
being returned if we ran out of retries.

This is related to lp:1604644 (duplicate key error still happening),
although the problem there is actually that the binaries running don't
have the mgo patch applied, compounded by lp:1605050 which will replace
patched binaries from upload tools with unpatched ones. (Finding the bug
in the patch meant that we could be sure that the binaries that raised
duplicate key errors were unpatched.)

Also adds logging so we can see clearly in the log when the patch has
been applied.