Add the machine removals collection to state
This will store information about machines that have been removed but
still need to be cleaned up in the provider.

Machine.Remove will add a record to machineRemovalsC with the
information to clean up, and a worker with access to the environ will
watch those records and clean them up.

Added State.AllMachineRemovals and .ClearMachineRemovals to manage them,
and a notify watcher to drive the worker.