state/backups: don't back up admin db
There's no evidence suggesting we want or need any of that information;
and when we *do* try to restore the admin DB, we frequently end up
nuking our own authorization to do exactly that: see lp:1604641. Have
not been able to reproduce such failures with this change.

This change is otherwise not tested, because the ignore-some-databases
code is so entirely incoherent. I've left notes in the code; if anyone
*can* imagine a test that has actual predictive power, please let me
know so that I can write it.