lxd: Properly store the image source (auto-update)
In current code, Juju resolves the image alias prior to copying the
image into the local image store.

This means that the local LXD daemon won't know what the image alias was
and so can't auto-update the image in the future.

Fix this by just passing the alias straight to LXD, but still resolve it
in Juju before so the log messages and errors remain identical.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>