state: fix unit/machine storage cleanup
There was a problem with the way we were cleaning up
storage when force-destroying machines:
 - we were destroying machine storage (volumes) before
   unit storage; it should be the other way around
 - due to holdover logic, we were not allowing machine-scoped
   storage to be short-circuit removed if it wasn't yet
   provisioned

The latter issue is due to an old issue which no longer
affects us: at one point, we leaked loop devices from
LXC containers, and so we marked them as "persistent".
Thus, we could not safely destroy a machine if it had
unprovisioned machine-scoped storage, or we would risk
leaking loop devices. This is no longer an issue, and
so we now assume machine-scoped storage will be destroyed
along with the machine. Thus, a machine can be removed
even if it has unprovisioned machine-scoped storage
associated with it.

Fixes https://bugs.launchpad.net/juju-core/+bug/1604931