{apiserver,api}/common/cloudspec: cloudspec API
Add a common CloudSpec API to various facades,
for workers and other clients that need to
construct an Environ. The controller facade
needs to embed this API for "destroy-controller"
to work without access to bootstrap config.

The API method is non-bulk, like ModelConfig,
as it doesn't make sense to specify model tags
in the facades that the API is included in.
In all cases, all methods in the facade are
scoped to a single model. In the case of the
controller, we only want to be able to query
the CloudSpec of the controller model.

These will be used in a following PR, where we
add CloudSpec to environs.OpenParams.

**QA**

Bootstrapped lxd.