Implement methods needed for Networking interface
The vsphere provider was being treated as not supporting networking,
despite having some support (for Subnets and NetworkInterfaces). Add
stub implementations for Spaces and SupportsSpaceDiscovery - they are
needed for environs.SupportsNetworking, otherwise Subnets and
NetworkInterfaces wouldn't be used.

Also add a test that will fail if new methods are added to the
Networking interface but not added here.

Fixes lp:1606917.