environs: add CloudSpec
Add the environs.CloudSpec type, which contains
the information required to connect to the cloud
provider. This includes the cloud type and name;
region name; endpoints; and credentials. This
will later be added to OpenParams, and also
PrepareForCreateEnvironment (or we may refactor
EnvironProvider.BootstrapConfig so that it is
used for both bootstrap and hosted models).

environs.BootstrapConfigParams and
environs/bootstrap.PrepareParams are both updated
to use the new struct instead of spreading the
contents out like they used to. The providers are
of course updated to match.

The state/stateenvirons package has been updated
with a method of producing a CloudSpec from a
*state.State. This will be used in a follow-up
PR that introduces a CloudSpec method to several
apiserver facades.