uniter/remotestate_test: use clock more carefully
The original tests were Advance()ing the clock without any reason to
assume the loop was actually waiting for it. Thanks to the complex
nature of the remote state watcher itself, precise alarm counting is
unlikely to be robust in the long term; but waiting for the alarms to
stop coming appears to be effective, and is certainly better than not
waiting *at all* as we did before.

fixes lp:1604955