cmd/jujud/agent: Reenable test is not affected by LP 1600301
As part of sorting out LP 1600301, TestDyingModelCleanedUp was
disabled but because it doesn't use TrackModels it is unlikely to have
the problem which the bug relates to.

Also, StartSync was being called on the wrong State. Calling StartSync
on a State for the correct model makes the test somewhat faster
(although I'm still not sure exactly how this works as it's not the
same State instance used by the apiserver).