Local charm revision metadata is once again saved
One of the things the fix for lp:1580418 did is remove obsolete charm documents from mongo. This had the side-effect of resetting a local charm's revision number when the charm was completely removed from the environment. There is another bug where the filesystem cache is not cleared out when a charm is removed, and fixing this bug would also fix this bug, but after discussing with William Reade, it was apparent that the more holistic approach was to bring local charms back into the fold with regards to a charm URL representing a specific revision of a charm archive. Breaking this concept can have many unintended consequences.

Fixes: https://bugs.launchpad.net/juju-core/+bug/1605096