cmd/jujud/agent: Re-enable TestHostedModelWorkers
The dummy provider blows up in the face of multi-model scenarios
causing TestHostedModelWorkers to fail.

The environ-tracker worker now takes a func to create a new Environ to
support injecting of an alternate Envrion in tests. A new minimal
Environ is now injected in TestHostedModelWorkers which is just enough
to allow the various model workers to run.

TestHostedModelWorkers is now works again and has been re-enabled.

This is the final PR for LP 1600301.