Simplify handling of machine removal in state
Letting a machine be removed and then doing cleanup sometime after has a
problem if that machine is the last one keeping the model alive. Rather
than doing that, mark the machine as ready for removal and have the
removal cleanup/completion remove the machine as well. This is simpler
and has the side benefit that the worker will have somewhere to store
status messages if there are problems performing some of the cleanup.