worker: less unreliable test
See lp:1538771 and comments in code.

Note that this is not *reliable*, just *less unreliable*; the long term
solution is to drop worker.Runner, or rewrite it to be observable; (or,
perhaps, rewrite it to render this usage mode impossible; but that's
most similar to just dropping it).

QA steps: figure out how to load your machine enough to trigger failure
under stress.bash; then apply the patch, run stress.bash, and load the
machine again. (I ran a couple of sister VMs and a couple of heavyish
games on the host.)