worker/migrationmaster: Always unlock fortress when no migration is active
The fortress was being left locked when there was a previously aborted
migration. This was preventing workers from restarting following an
aborted migration.

This change also fixes a potentially racy test
(TestPreviouslyCompletedMigration). It was assuming that all API calls
would be made before the worker was killed and there are no guarantees
of that.