Add GetMachineProviderInterfaceInfo
This takes a set of machine tags and returns the information needed to
talk to the provider about the interfaces attached to the machines.

This uses state.Machine.AllProviderInterfaceInfos (which returns simple
objects rather than entities that embed a *State), so we can get rid of
all of the complicated shims and mocks for State -> Machine ->
LinkLayerDevice -> Address -> Subnet.