Manual machines shouldn't stop destroy-controller
For a while now we've been ref-counting machines
against models, and so we will no longer leak manual
machines when running "juju destroy-controller" or
"juju destroy-model". Thus, we drop the check for
manual machines in model destruction, and rely on
the undertaker and cleanups to remove them manual
machines.

The manual provider's Destroy method is now a no-op,
and the existing functionality is only performed in
the DestroyController method.

Fixes https://bugs.launchpad.net/juju-core/+bug/1475212
Fixes https://bugs.launchpad.net/juju-core/+bug/1611067