Pass ReleaseContainerAddresses much less info
Fully constructing a network.InterfaceInfo is way more than we need to
actually release the container addresses, and it's a hassle to build it
in state and pass through the API, especially with testing. But it seems
wrong/confusing to only set the MAC address and leave the rest blank
because we know the MAAS provider only needs the MAC.

Instead of that, create a struct that is designed to hold information
for identifying an interface/L2 device to a provider (which is a much
smaller set of fields) and always fully populate that information. When
we need other fields for other providers we can add it.

Part of lp:1585878.

Only unit testing done - the worker that will call
ReleaseContainerAddresses is still under construction.