Update charmstore and charmrepo dependencies.
This also allow us to introduce support for the new charm store channels in Juju.
Long story short, the charm store is now ready to expose the new channels model based on 4 channels (stable, candidate, beta and edge), so that it is in line with the snap store.
There is no specific bug for this, but this is a requirement for the new charm store, and the next release will support this: FYI in this process all development charms and bundles will be migrated to the edge channel.

NOTE: we want to include this in the next beta, so that we can release the new charm store as soon as possible.
When cs will be released, all previous beta client will be broken as they will be no longer able to deploy charms in most cases: we'll have to invite users to upgrade to the newest beta.
This is due to the panic in this piece of old code: https://github.com/juju/charmrepo/blob/f238e1ad216e60450f2eb2eebfedb57237664ce6/charmstore.go#L273