rpc: rename MethodFinder to Root; add Killer
This makes it so all API root implementations must implement
Killer, so it's much harder to accidentally remove support
for tearing down API calls.

We rename the interface because it's not just a method
finder any more, and Root fits well with the variable
naming used in lots of places.

We also remove rpc.Cleaner, as it's not used any more and
would clutter the Root interface for no good reason. It's
not clear that it was something that was that useful anyway.