api/controller: add failing test for API server termination
Currently the API server cannot be terminated if there is an
outstanding Next call on a controller-only connection.

This PR adds a (failing, therefore skipped) test for that,
which will be fixed in a subsequent PR.

As part of doing this, we add support for tests to obtain controller-only
API connections.