apiserver: serve controller-only endpoint from /api
This makes the websocket API endpoints more consistent
between controller-only and model. We continue to
server to old endpoint at / for backward compatibility.

We also change the API server so that it checks that
the login result is consistent with the model UUID
being asked for, and store the model UUID in parsed
form so we don't parse it each time it's asked for.

This means we can return a boolean rather than an error
from the ModelTag entry point, which prepares the way for
all facades to check whether they're being layered onto
the right kind of API connection.

We also clear up some backward-compatibility cruft
from the apiserver implementation.