api/modelmanager: introduce DestroyModels
Introduce a new API method, DestroyModels, which
supersedes DestroyModel. The new method takes a
set of model tags, and returns a set of error
results. The old method will be dropped once the
GUI and other clients are updated.

Drive-by: removed ModelInfo from the client-side
of the Client API. I'll notify the list about it
being deprecated, and remove from the server side
soon.