environs/sync: Make TestMockBuildTools reliable
This test was relying on the specifics's of Go's gzip and tar
implementation. Even for a given Go version the compression output can
vary sometimes and between Go versions it's more likely to be
different. The test was failing consistently under Go 1.7.

Instead of comparing sizes and checksums, the mock tools tarball is
extracted to ensure the contents inside are correct.

Fixes LP 1613200.