Use CloudCredentialTag
We update the Cloud and ModelManager APIs to
make use of the new names.CloudCredentialTag
type. This decouples credentials from model
owner in the CreateModel API call, such that
we can later add ACLs on, and share, credentials.

The Cloud.CloudCredentials API now returns
only the *tags* of cloud credentials that a
user has access to, rather than the contents
of those credentials. This will make it possible
to share cloud credentials with another user
without exposing the credential contents to them.

Note that ACLs and sharing are *not* implemented.
When you update credentials, you're still limited
to updating your own. For now, you can still only
refer to your own, even if the tag syntax allows
for referring to another user's.