Do not allow register of alternative controller names
The register command should not allow a controller to be added
more than once, with alternative names, to a controller collection.

jujuclient/file.go now implements AddController in addition to
UpdateController, always using AddController to add, and UpdateController
to update. Update had previously updated any existing controller of the
same name in the local store, and otherwise added a controller as new,
even if it was effectively an alias of an existing controller by UUID.

juju/testing/conn.go JujuConnSuite can now create additional user home
environments within a test. This allows `add-user` and `register` to be
tested without controller UUID contention.

Fixes: #1593350